requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
future==0.17.1            # via python-twitter
idna==2.8                 # via requests
oauthlib==3.0.1           # via requests-oauthlib
python-twitter==3.5
requests-oauthlib==1.2.0  # via python-twitter
requests==2.21.0          # via python-twitter, requests-oauthlib
selenium==3.141.0
urllib3==1.24.1           # via requests, selenium
